+++ docs/reference/changelog.rst
@@ -56,6 +56,8 @@ Bug fixes:
 - Set ``DEBIAN_FRONTEND=noninteractive`` in the global execution environment,
   so all steps run non-interactively by default.
 - Add git source mutual exclusivity to JSON schema.
+- Update the ``uv``, ``poetry``, and ``npm=`` plugins to use native TLS
+  certificates.
 
 .. _release-2.34.1: